+++ src/test/java/org/geowe/service/geometry/test/IntersectionTest.java
@@ -43,6 +43,7 @@ public class IntersectionTest {
 	private ResteasyClient restClient;
 	private ResteasyWebTarget target;
 	private static final String INTERSECTION_WKT = "POLYGON ((-4.776890766405977 37.59953857604632, -4.783158473202792 37.60085706996653, -4.768746321046889 37.63347410816212, -4.758777428165223 37.63424429972881, -4.75776057778995 37.63619261691262, -4.72050423716731 37.637201266895275, -4.712456824160618 37.63782300590933, -4.712404981850047 37.63742054000688, -4.67743357013068 37.63836732955289, -4.6410461453507175 37.597580901219175, -4.630749610620046 37.55351410381898, -4.63144176324856 37.553438549085996, -4.72549216536138 37.54317209523422, -4.787279223379428 37.579633966944684, -4.776890766405977 37.59953857604632))";
+	private static final double TOLERANCE = -0.00001;
 	@Before
 	public void setUp() throws Exception {
 		restClient = new ResteasyClientBuilder().build();
@@ -54,7 +55,7 @@ public class IntersectionTest {
 	public void intersectionTest() {
 		JtsIntersectionService service = new JtsIntersectionService();
 		Response response = service.getIntersection(DataTestProvider
-				.getIntersectionData());
+				.getIntersectionData(), TOLERANCE);
 		FlatGeometry intersectionGeometry = (FlatGeometry) response.getEntity();
 
 		Assert.isTrue(response.getStatus() == Status.CREATED.getStatusCode(),
@@ -80,7 +81,7 @@ public class IntersectionTest {
 	@Test
 	public void intersectinPostNullbodyTest() {
 
-		Response response = target.request().post(
+		Response response = target.queryParam("tolerance", TOLERANCE).request().post(
 				Entity.entity(null, "application/json;charset=UTF-8"));
 
 		ErrorEntity error = response.readEntity(ErrorEntity.class);