+++ CHANGELOG.md
@@ -9,6 +9,12 @@ You can see the previous changelog [here] (http://www.geowe.org/index.php?id=cha
 - Geometry validations
 - Asynchronous tasks
 
+## [0.1.0]SNAPSHOT - 2017-03-08
+
+### Added
+- Implemented resources for JTS operations:
+	- Difference
+	- Symetric Difference 
 
 ## [0.0.1]SNAPSHOT - 2017-03-06